Add ‘documentation’ to described components
Add a property `documentation` to described components. It returns an object containing `toJSON`, `toMarkdown`, and `toTypescript` functions, identical to the ones defined on the described component itself.

The purpose of this is to make a more readable API, so that someone unaware of `react-desc` can see that it is related to documentation.